editor path functionality cleanup
