+++ src/event/display/EditorChangeEvent.java
@@ -1,0 +1,6 @@
+package event.display;
+
+import event.editor.EditorEvent;
+
+public class EditorChangeEvent extends EditorEvent {
+}
