+++ src/event/editor/MapResizeEvent.java
@@ -1,0 +1,4 @@
+package event.editor;
+
+public class MapResizeEvent extends EditorEvent{
+}
